Fixed a floating-point number bug
The variable "isScientific", which signifies that the current token is a scientific-notation numeric, was not reset to false after adding the token to the token list.  This caused line 58 to return true, and then fire an exception on line 59, if another scientific notation numeric was encountered:

 if (isScientific && IsScientificNotation(characters[i]))
        throw new ParseException(string.Format("Invalid token \"{0}\" detected at position {1}.", characters[i], i));

Reseting "isScientific" after every scientific numeric token is added eliminates this bug.